Fct 'check_time_consistency': modified + moved in 'classes.py'
