+++ doc/functions_utilities.rst
@@ -33,6 +33,6 @@ csync
 check_time_consistency: check time consistency of first variable of a dataset or an ensemble
 ----------------------------------------------------------------------------------------------
 
-.. autofunction:: climaf.driver.check_time_consistency
+.. autofunction:: climaf.classes.check_time_consistency
